refactor: :zap: create mdsvex.config.js
move mdsvex configuration to a dedicated file; modify the default options to disable fancy typography, point to a named layout, use a custom version of two rehype plugins, highlight code fences with shiki
