+++ src/routes/blog/snake-svg-svelte.md
@@ -1,7 +1,7 @@
 ---
 title: Snake, SVG, Svelte
-date: 2022-07-16T20:53
-brief: Let's build a game of Snake with vector graphics and a rather popular utility.
+description: Let's build a game of Snake with vector graphics and a rather popular utility.
+datetime: '2022-07-16T20:53'
 ---
 
 Snake is a game which needs very little introduction, except perhaps a demo.
