+++ svelte.config.js
@@ -1,28 +1,7 @@
 import adapter from '@sveltejs/adapter-auto';
 import { resolve } from 'path';
-import { mdsvex, escapeSvelte } from 'mdsvex';
-import slug from 'rehype-slug';
-import autolinkHeadings from 'rehype-autolink-headings';
-import { getHighlighter } from 'shiki';
-
-const highlighter = async (code, lang) => {
-	const shikiHighlighter = await getHighlighter({ theme: 'material-palenight' });
-	const shikiCode = shikiHighlighter.codeToHtml(code, { lang });
-
-	return escapeSvelte(shikiCode);
-};
-
-const mdsvexConfig = {
-	extensions: ['.md', '.svx'],
-	smartypants: false,
-	layout: {
-		blog: 'src/lib/layout/Blog.svelte'
-	},
-	rehypePlugins: [slug, autolinkHeadings],
-	highlight: {
-		highlighter
-	}
-};
+import { mdsvex } from 'mdsvex';
+import mdsvexConfig from './mdsvex.config.js';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
